pyproject.toml
[project]
name = "playground-model-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ai-server-sdk>=0.0.27",
    "openai>=2.2.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
